moved light structs to a single header file
