+++ UniversoEngine/src/renderer/Lighting.cpp
@@ -1,0 +1,1 @@
+#include "Lighting.h"
